# openapi.yaml
openapi: 3.0.3
info:
  title: 焚信 BurnMsg API
  version: 2.0.0
  description: |-
    焚信（BurnMsg）是一款企业内部加密即时通讯（IM）应用，支持云聊天与秘密聊天、阅后即焚消息。本文件为其后端（NestJS + TypeORM + MySQL）的标准 OpenAPI 3.0.3 接口文档。

    【账号体系】
    - 不开放自助注册，所有账号由管理后台统一开通（POST /accounts，仅 admin）。
    - 用户名即公司内部手机号（1 开头 11 位数字）。
    - 新开通 / 被重置密码的账号 force_change_pwd=true，前端应强制跳转修改密码，修改成功后该标记置为 false。

    【认证方式】
    - 受保护接口在请求头携带：Authorization: Bearer <access_token>。
    - 双 token 机制：access_token（JWT_SECRET 签名，默认有效期 7200 秒）+ refresh_token（独立的 JWT_REFRESH_SECRET 签名，默认有效期 2592000 秒 / 30 天，payload 带 type=refresh 标记防混用）。
    - access_token 过期后用 refresh_token 调 POST /auth/refresh-token 换取新 token 对。
    - 即时吊销：JWT 守卫除验签外，每个受保护请求都会实时查库校验「账号仍为 active」且「签发 token 的设备记录仍存在」；账号被停用或设备被下线后，已签发的 token 立即失效（全端即时强制下线）。

    【统一响应格式】
    - 成功：{ "code": 0, "message": "ok 或各接口自定义提示语", "data": <业务数据，无内容时为 null> }
    - 失败：{ "code": <HTTP 状态码>, "message": "<错误描述；参数校验失败时为多条错误以 '; ' 拼接>" }

    【通用约定】
    - 分页参数：page（默认 1）、pageSize（默认 20，最大 200；通讯录列表默认 50）。
    - 请求中的时间字段使用 ISO8601 字符串；响应中的时间为服务端 datetime。
    - 所有业务 ID 均为 UUID v4 字符串。
    - 角色：user / admin。标注「仅管理员」的接口仅 admin 可调用，普通用户返回 403。
    - 强制改密拦截：force_change_pwd=true 的账号（新开通 / 被重置密码后），除 POST /auth/change-password 与 GET /auth/profile 外，调用其他任何受保护接口一律返回 403（FORCE_CHANGE_PASSWORD），前端应引导其先完成改密再进入主界面。
    - 软删除账号：被 DELETE /accounts/{id} 软删的账号立即吊销全部 token 并强制下线，其历史消息保留、前端统一展示「已注销用户」；账号列表默认不含已软删账号，加 show_deleted=1 可查看。
servers:
  - url: http://localhost:9091/api/v1
...
security:
  - bearerAuth: []
...
paths:
...
  /announcements:
    post:
      tags: [公告]
      summary: 发布系统公告（仅 admin）
      description: |-
        【用途】管理员发布一条系统公告，分 normal / urgent 两级。

        【调用时机】管理后台「系统公告」发布操作时。

        【业务规则与注意事项】
        - 仅 admin 角色可调用。
        - normal：进入 App 公告中心，用户打开公告页时拉取；urgent：除进公告中心外，还通过 WebSocket 实时推送 announcement:new 事件给在线用户（前端弹窗）。
        - target_type=department 时 target_departments 必填，部门名需与账号 department 字段完全一致，否则目标用户不可见。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, content]
              properties:
                title:
                  type: string
                  maxLength: 200
                  description: 公告标题
                  example: 系统维护通知
                content:
                  type: string
                  description: 公告正文
                  example: 本周六 22:00-24:00 服务器升级维护，期间消息可能延迟。
                priority:
                  type: string
                  enum: [normal, urgent]
                  default: normal
                  description: urgent 会走 WebSocket 实时弹窗推送到在线 App
                target_type:
                  type: string
                  enum: [all, department]
                  default: all
                  description: all=全员；department=按部门定向
                target_departments:
                  type: array
                  items: { type: string }
                  description: target_type=department 时必填，部门名单（需与账号部门完全一致）
                  example: ['技术部', '市场部']
      responses:
        '200':
          description: 发布成功（message 为「公告发布成功」），data 为已入库的公告对象
          content:
            application/json:
              schema:
                type: object
                properties:
                  code: { type: integer, example: 0 }
                  message: { type: string, example: 公告发布成功 }
                  data:
                    $ref: '#/components/schemas/Announcement'
        '400':
          description: 参数校验失败（如 target_type=department 但未传部门名单）
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '401':
          description: 未携带或无效/过期 token；账号已停用或设备已下线
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '403':
          description: 非 admin 角色（Admin access required）
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
    get:
      tags: [公告]
      summary: App 端公告列表（公告中心）
      description: |-
        【用途】当前用户拉取自己可见的公告列表（全员公告 + 定向到本部门的公告）。

        【调用时机】App 公告中心页加载/下拉刷新时。

        【业务规则与注意事项】
        - 按发布时间倒序；每条附 is_read 标记（是否已读），未读的可配合未读数角标展示。
        - 部门定向公告仅目标部门成员可见。
      parameters:
        - name: page
          in: query
          schema: { type: integer, minimum: 1, default: 1 }
          description: 页码，默认 1
        - name: pageSize
          in: query
          schema: { type: integer, minimum: 1, maximum: 200, default: 20 }
          description: 每页条数，默认 20
      responses:
        '200':
          description: '成功，data 为 { data: 公告数组（含 is_read）, total }'
          content:
            application/json:
              schema:
                type: object
                properties:
                  code: { type: integer, example: 0 }
                  message: { type: string, example: ok }
                  data:
                    type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Announcement'
                      total: { type: integer, example: 12 }
        '401':
          description: 未携带或无效/过期 token；账号已停用或设备已下线
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /announcements/manage:
    get:
      tags: [公告]
      summary: 公告管理列表（仅 admin，含已读人数）
      description: |-
        【用途】管理员查看全部公告及每条的已读人数，用于追踪触达效果。

        【调用时机】管理后台「系统公告」列表页。

        【业务规则与注意事项】
        - 仅 admin 角色可调用；含全量公告（不分部门可见性），每条附 read_count（已读人数）。
      parameters:
        - name: page
          in: query
          schema: { type: integer, minimum: 1, default: 1 }
          description: 页码，默认 1
        - name: pageSize
          in: query
          schema: { type: integer, minimum: 1, maximum: 200, default: 20 }
          description: 每页条数，默认 20
      responses:
        '200':
          description: '成功，data 为 { data: 公告数组（含 read_count）, total }'
          content:
            application/json:
              schema:
                type: object
                properties:
                  code: { type: integer, example: 0 }
                  message: { type: string, example: ok }
                  data:
                    type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Announcement'
                      total: { type: integer, example: 12 }
        '401':
          description: 未携带或无效/过期 token；账号已停用或设备已下线
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '403':
          description: 非 admin 角色（Admin access required）
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /announcements/unread-count:
    get:
      tags: [公告]
      summary: 当前用户未读公告数
      description: |-
        【用途】返回当前用户可见公告中未读的条数，用于 App 公告中心入口角标。

        【调用时机】App 启动/回到前台/收到 announcement:new 推送后刷新角标时。
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  code: { type: integer, example: 0 }
                  message: { type: string, example: ok }
                  data:
                    type: object
                    properties:
                      unread: { type: integer, description: 未读公告条数, example: 3 }
        '401':
          description: 未携带或无效/过期 token；账号已停用或设备已下线
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /announcements/{id}:
    delete:
      tags: [公告]
      summary: 删除公告（仅 admin）
      description: |-
        【用途】删除一条公告，同时清除其全部已读记录；已推送的弹窗不撤回，公告中心列表不再展示。

        【调用时机】管理后台「系统公告」对过期/误发公告执行删除时。
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
          description: 公告 ID（UUID）
      responses:
        '200':
          description: 删除成功（message 为「公告已删除」），data 为 null
          content:
            application/json:
              schema:
                type: object
                properties:
                  code: { type: integer, example: 0 }
                  message: { type: string, example: 公告已删除 }
                  data: { nullable: true, example: null }
        '401':
          description: 未携带或无效/过期 token；账号已停用或设备已下线
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '403':
          description: 非 admin 角色（Admin access required）
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '404':
          description: 公告不存在
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /announcements/{id}/read:
    post:
      tags: [公告]
      summary: 标记公告已读
      description: |-
        【用途】当前用户将一条公告标记为已读；一人一公告一条已读记录（唯一约束），重复调用幂等。

        【调用时机】用户点开公告详情时。
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
          description: 公告 ID（UUID）
      responses:
        '200':
          description: 成功（message 为「已标记为已读」），data 为 null
          content:
            application/json:
              schema:
                type: object
                properties:
                  code: { type: integer, example: 0 }
                  message: { type: string, example: 已标记为已读 }
                  data: { nullable: true, example: null }
        '401':
          description: 未携带或无效/过期 token；账号已停用或设备已下线
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '404':
          description: 公告不存在
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |-
        JWT access token。登录（POST /auth/login）或刷新（POST /auth/refresh-token）后获得，放入请求头：
        Authorization: Bearer <access_token>
        默认有效期 7200 秒；每个受保护请求都会实时校验账号状态与设备存在性，账号停用或设备下线后 token 立即失效。
  schemas:
    Error:
      type: object
      description: 统一错误响应（code 为 HTTP 状态码；参数校验失败时 message 为多条错误以 '; ' 拼接）
      properties:
        code:
          type: integer
          description: HTTP 状态码（400/401/403/404/409/413/429/500）
          example: 400
        message:
          type: string
          description: 错误描述
          example: 手机号格式不正确
      required: [code, message]
...
    Announcement:
      type: object
      description: 系统公告对象（App 端列表附 is_read；管理列表附 read_count）
      properties:
        id:
          type: string
          format: uuid
          description: 公告 ID（UUID v4）
        title:
          type: string
          description: 公告标题
          example: 系统维护通知
        content:
          type: string
          description: 公告正文
          example: 本周六 22:00-24:00 服务器升级维护，期间消息可能延迟。
        priority:
          type: string
          enum: [normal, urgent]
          description: normal 进公告中心；urgent 另走 WS 实时弹窗（announcement:new）
          example: normal
        target_type:
          type: string
          enum: [all, department]
          description: all=全员；department=按部门定向
          example: all
        target_departments:
          type: array
          items: { type: string }
          nullable: true
          description: 定向部门名单（target_type=department 时有值）
          example: null
        created_by:
          type: string
          format: uuid
          description: 发布人（管理员）用户 ID
        is_read:
          type: boolean
          description: 当前用户是否已读（仅 App 端列表接口返回）
          example: false
        read_count:
          type: integer
          description: 已读人数（仅管理列表 GET /announcements/manage 返回）
          example: 35
        created_at:
          type: string
          format: date-time
          example: '2026-08-27T08:00:00.000Z'
        updated_at:
          type: string
          format: date-time
          example: '2026-08-27T08:00:00.000Z'
...